move resetDemobutton
